fix warnings from Wall and use canonical functions

+++ lib/include/fimpera-lib/CBF.hpp
@@ -15,13 +15,13 @@ class CBF {
     uint64_t _nbCells;
     uint64_t _limitValueInBucket;
     uint64_t getIndex(const std::string& kmer) const;
-    int get(const uint64_t& index) const;
+    uint64_t get(const uint64_t& index) const;
 
    public:
     CBF(uint64_t nbBits = 100, uint64_t nbBuckets = 1);
     CBF(std::ifstream& fin);
-    int set(const std::string& kmer, int occurrence = 1);
-    int get(const std::string& kmer) const;
+    uint64_t set(const std::string& kmer, uint64_t occurrence = 1);
+    uint64_t get(const std::string& kmer) const;
     void dump(std::ofstream& fout) const;
     bool operator==(const CBF& that) const;
     ~CBF();
